Bugfix in scale_y_discrete() call in plot_all_metrics.R.
For some reason, ggplot2::expansion() could not be found when running in snakespeare virtual environment on Windows. This type of thing is literally what the virtual environment is supposed to prevent so I have no idea what's happening...

ggplot2::expansion() is a convenience function so I replaced the call with the result. We'll see if that fixes the issue.