feat: get_dialogue.sh now removes 2-line stage directions
As far as I have observed, whenever a stage direction starts in the middle of a line and continues to the next line, it finishes in the next line. If the stage direction is 2 full lines or more, the Folger formatting rules will insert a line break before the stage direction.

So, this commit takes care of the only remaining case where a stage direction starts in the middle of a line and ends in the following line.