fix: Fix sneaky yet disastrous regex issue in get_dialogue.sh
Fixed regex bug that caused partial line stage directions not preceded by a space (i.e., starting at the beginning of the line) to be missed, then caught by the multi-line stage direction removal, which ended up cutting all dialogue until the end of the NEXT stage direction. This caused large chunks of the play to be removed and not counted for dialogue, most noticably in Hamlet.